Resolved signup route's dropping names, added no match page
