+++ client/src/components/Meetings/CurrentMeeting/Subcomponents/index.js
@@ -1,0 +1,2 @@
+export { default } from "./Tableheader";
+export { default } from "./Tablerow";
