+++ client/src/components/Meetings/CurrentMeeting/index.js
@@ -1,0 +1,1 @@
+export { default } from "./CurrentMeeting";
