+++ client/src/components/Meetings/MeetingDetail/index.js
@@ -1,0 +1,1 @@
+export { default } from "./MeetingDetail";
