+++ client/src/components/Meetings/MeetingStats/index.js
@@ -1,0 +1,1 @@
+export { default } from "./MeetingStats";
